remove unnecessary function and move the start_workfile_template_builder to pipeline.py
