+++ client/ayon_nuke/api/lib.py
@@ -2734,24 +2734,6 @@ def prompt_reset_context():
         dialog.deleteLater()
 
 
-def start_workfile_template_builder() -> None:
-    """Trigger workfile template builder when a new file is created."""
-    if current_file() is not None:
-        return
-
-    from .workfile_template_builder import (
-        trigger_on_app_launch,
-        trigger_on_new_file,
-    )
-    if not os.getenv("AYON_NUKE_INITIALIZED"):
-        log.info("Triggering workfile template builder on application launch.")
-        trigger_on_app_launch()
-        os.environ["AYON_NUKE_INITIALIZED"] = "True"
-    else:
-        log.info("Triggering workfile template builder on new file.")
-        trigger_on_new_file()
-
-
 def add_scripts_menu():
     try:
         from scriptsmenu import launchfornuke
